backend/requirements.txt
django
djangorestframework
psycopg2
PyJWT